requirements.txt
joblib==1.0.1
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.0
scikit-learn==0.24.1
scipy==1.6.1
threadpoolctl==2.1.0
torch==1.7.1
torchvision==0.8.2
typing-extensions==3.7.4.3
